requirements.txt
typer
openai
python-dotenv
google-generativeai
